test: add import prefix and inheritance subclass
