--- xml_serializable/test/test_class_2.dart
@@ -1,9 +1,0 @@
-import 'package:xml_annotation/xml_annotation.dart';
-
-@XmlSerializable()
-class TestClass {
-  @XmlElement()
-  String? fooBar;
-
-  TestClass({this.fooBar});
-}
